CRM-11045 - Add CRM_Extension_Info::loadFromString()

+++ tests/phpunit/CRM/Extension/InfoTest.php
@@ -5,15 +5,18 @@ require_once 'CiviTest/CiviUnitTestCase.php';
 class CRM_Extension_InfoTest extends CiviUnitTestCase {
   function setUp() {
     parent::setUp();
-    $this->file = tempnam(sys_get_temp_dir(), 'infoxml-');
+    $this->file = NULL;
   }
 
   function tearDown() {
-    unlink($this->file);
+    if ($this->file) {
+      unlink($this->file);
+    }
     parent::tearDown();
   }
 
-  function testGood() {
+  function testGood_file() {
+    $this->file = tempnam(sys_get_temp_dir(), 'infoxml-');
     file_put_contents($this->file, "<extension key='test.foo' type='module'><file>foo</file><typeInfo><extra>zamboni</extra></typeInfo></extension>");
 
     $info = CRM_Extension_Info::loadFromFile($this->file);
@@ -22,13 +25,36 @@ class CRM_Extension_InfoTest extends CiviUnitTestCase {
     $this->assertEquals('zamboni', $info->typeInfo['extra']);
   }
 
-  function testBad() {
+  function testBad_file() {
     // <file> vs file>
+    $this->file = tempnam(sys_get_temp_dir(), 'infoxml-');
     file_put_contents($this->file, "<extension key='test.foo' type='module'>file>foo</file></extension>");
 
     $exc = NULL;
     try {
       $info = CRM_Extension_Info::loadFromFile($this->file);
+    } catch (CRM_Extension_Exception $e) {
+      $exc = $e;
+    }
+    $this->assertTrue(is_object($exc));
+  }
+
+  function testGood_string() {
+    $data = "<extension key='test.foo' type='module'><file>foo</file><typeInfo><extra>zamboni</extra></typeInfo></extension>";
+
+    $info = CRM_Extension_Info::loadFromString($data);
+    $this->assertEquals('test.foo', $info->key);
+    $this->assertEquals('foo', $info->file);
+    $this->assertEquals('zamboni', $info->typeInfo['extra']);
+  }
+
+  function testBad_string() {
+    // <file> vs file>
+    $data = "<extension key='test.foo' type='module'>file>foo</file></extension>";
+
+    $exc = NULL;
+    try {
+      $info = CRM_Extension_Info::loadFromString($data);
     } catch (CRM_Extension_Exception $e) {
       $exc = $e;
     }
